I have added the navigation menu and the header section for the website.
I have added the navigation menu and the header section for the website.

Task to be done on this page is to make the header section background image responsive and link all the other pages. i.e services, contact us and about us page.  And the footer section also have to uploaded.